+++ src/features/dev-panel/DevPanel.tsx
@@ -3,49 +3,12 @@ import { createPortal } from 'react-dom';
 import { useDevBridge } from './useDevBridge';
 import { ToastContainer } from './components/Toast';
 import { FileText, Users, Image, X, UserCog, WifiOff, Loader2, AlertCircle, Globe } from 'lucide-react';
-import { makeTokens } from '@/shared/tokens/theme';
+import { ThemeTokensContext, makeT, type TTokens, useIsDark } from './theme';
 
 const DocsPanel     = lazy(() => import('./panels/DocsPanel'));
 const ContactsPanel = lazy(() => import('./panels/ContactsPanel'));
 const AssetsPanel   = lazy(() => import('./panels/AssetsPanel'));
 const SitePanel     = lazy(() => import('./panels/SitePanel'));
-
-// ─── Тема ────────────────────────────────────────────────────────────────────
-
-export function useIsDark(): boolean {
-  const [isDark, setIsDark] = useState(() =>
-    typeof document === 'undefined'
-      ? true
-      : document.documentElement.classList.contains('dark')
-  );
-  useEffect(() => {
-    const onTheme   = (e: Event) =>
-      setIsDark((e as CustomEvent<{ isDark: boolean }>).detail.isDark);
-    const onStorage = (e: StorageEvent) => {
-      if (e.key === 'theme') setIsDark(e.newValue !== 'light');
-    };
-    globalThis.addEventListener('hub:theme-change', onTheme);
-    globalThis.addEventListener('storage', onStorage);
-    return () => {
-      globalThis.removeEventListener('hub:theme-change', onTheme);
-      globalThis.removeEventListener('storage', onStorage);
-    };
-  }, []);
-  return isDark;
-}
-
-export function makeT(isDark: boolean) {
-  const t = makeTokens(isDark);
-  return {
-    ...t,
-    mono: 'ui-monospace, "Cascadia Code", "Fira Code", monospace',
-    inpBorder: t.inpBdr,
-    editorFg: isDark ? '#e2e8f0' : '#1e293b',
-  };
-}
-
-export type TTokens = ReturnType<typeof makeT>;
-export const ThemeTokensContext = React.createContext<TTokens>(makeT(true));
 
 // ─── Табы ────────────────────────────────────────────────────────────────────
