+++ src/features/dev-panel/components/uiTheme.ts
@@ -1,0 +1,5 @@
+export const T = {
+  fg: 'rgba(255,255,255,0.9)',
+  fgSub: 'rgba(255,255,255,0.22)',
+  mono: 'ui-monospace, "Cascadia Code", "Fira Code", monospace',
+};
