+++ src/features/dev-panel/panels/SitePanel.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useCallback, useContext, useRef } from 'react';
 import { bridge } from '../useDevBridge';
-import { ThemeTokensContext } from '../DevPanel';
-import { toast } from '../components/Toast';
+import { ThemeTokensContext } from '../theme';
+import { toast } from '../components/toastBus';
 import { Loader2, LayoutTemplate, FileText, RefreshCw, AlertCircle } from 'lucide-react';
 import type { SiteConfig } from '../useDevBridge';
 import {
