+++ src/shared/contexts/useTheme.ts
@@ -1,0 +1,9 @@
+import { useContext } from 'react';
+import { ThemeContext } from './themeContextStore';
+import type { ThemeContextType } from './themeContextStore';
+
+export const useTheme = (): ThemeContextType => {
+  const context = useContext(ThemeContext);
+  if (!context) throw new Error('useTheme must be used within ThemeProvider');
+  return context;
+};
